harden: close H-03/04/05/07/13/14/15/16/31/39 + threat model
Configuration and policy gates from the use-protection-please audit,
each verified rather than assumed.

  H-03  rust-toolchain.toml pins 1.98.0 + rustfmt/clippy/llvm-tools.
  H-04  .cargo/config.toml gains frame pointers, and per-OS linker
        hardening (RELRO/now/noexecstack/PIE on Linux).
  H-05  overflow-checks = true and an explicit panic = "unwind" (the
        EDC worker's catch_unwind cleanup needs unwinding). The whole
        suite and 68/68 byte-identity pass WITH overflow checks on.
  H-07  Cargo.lock un-gitignored and tracked.
  H-08  deny.toml: advisories + licenses + bans + sources.
  H-09  RUSTSEC-2024-0384 ignored with a dated, reasoned entry -- it is
        dev-only via minifb and cargo tree --edges normal shows 0 hits.
  H-13  [sources] denies unknown registries and any git dependency.
  H-14  dependabot, grouped monthly, security updates ungrouped.
  H-15  workspace lint policy + clippy.toml, and CLEAN under -D warnings.
  H-16  all 17 undocumented unsafe blocks now carry SAFETY comments.
  H-31  control-flow-guard: the PE goes 0x8160 -> 0xc160, so ASLR,
        high-entropy ASLR, DEP and CFG are all set.
  H-01/02  docs/threat-model.md: a STRIDE pass naming assets,
        adversaries, the highest-value attack path and what is
        deliberately not defended.
  H-39  SECURITY.md with contact, response windows and a 90-day policy.

The lint policy is security-focused, not pedantic: unwrap/expect/panic
are DENIED on the crates that eat untrusted bytes and allowed on the
encoder (standard tier -- its input is the caller's own frames), and
undocumented_unsafe_blocks keeps H-16 from regressing. Every allow
carries its reason inline. 87 now-redundant local #[allow] attributes
removed, so the workspace policy is the single source.

Genuine defects the policy found, all fixed:
  * rust-version claimed 1.80 but the code uses is_none_or (1.82+), so
    a consumer on 1.80 got a compile error instead of a clear message.
  * two unwraps in rusty_h264-common's deblock (not previously audited).
  * dead arithmetic in two dev tools (x - x, and a * 0 term).
  * two strided nzc copies replaced with slice copies.

MEASURED COST of the mitigations, deterministic instruction census:
  baseline 285,897 -> +CFG 310,230 (+8.5%) -> +frame pointers 311,457
  (+0.4%). The split inverts the usual assumption: frame pointers are
  nearly free and CFG is the whole cost. Static count only -- CFG
  instruments indirect calls and the hot loops are direct or inlined,
  so treat 8.5% as a ceiling. Recorded in .cargo/config.toml.

Gates: 68/68 ALL BYTE-IDENTICAL; full workspace tests green; clippy
clean under -D warnings.

+++ crates/rusty_h264-accel/src/hpel.rs
@@ -164,7 +164,7 @@ unsafe fn hpel_fused_avx2(f: &[u8], pw: usize, ph: usize, h: &mut [u8], v: &mut 
         while x + 16 + 3 <= pw {
             let ld = |off: isize| {
                 _mm256_cvtepu8_epi16(_mm_loadu_si128(
-                    f.as_ptr().add((y0 + x) as usize).offset(off) as *const __m128i
+                    f.as_ptr().add(y0 + x).offset(off) as *const __m128i
                 ))
             };
             let (a, b, cc, d, e, g) = (ld(-2), ld(-1), ld(0), ld(1), ld(2), ld(3));
