+++ crates/rusty_h264-encoder/tests/cabac_roundtrip.rs
@@ -1,3 +1,11 @@
+// Dev/test target, not shipped: a panic here IS the diagnostic (that is what an
+// assertion is). The workspace's unwrap/expect/panic denials exist to keep them
+// off the decoder's untrusted-input path, so they are relaxed for this file.
+#![allow(clippy::unwrap_used, clippy::expect_used, clippy::panic)]
+// Dev tools also accumulate fields and helpers kept for the NEXT investigation;
+// dead code here is a scratchpad, not a defect.
+#![allow(dead_code, unused)]
+#![allow(clippy::unnecessary_unwrap, clippy::zombie_processes)]
 //! CABAC I-slice conformance gate (the bringup-encoder verification spine).
 //!
 //! For each QP × content, encode an all-intra clip TWICE — once CABAC, once CAVLC —
@@ -369,7 +377,7 @@ fn weightp_fade_roundtrips_and_wins() {
             let mut fr = split_frame(w, h, 0); // STATIC content...
             for p in fr.y.iter_mut() {
                 // ...under a strong global fade: -14 luma per frame.
-                *p = (*p as i32 - 14 * f as i32).clamp(16, 235) as u8;
+                *p = (*p as i32 - 14 * f).clamp(16, 235) as u8;
             }
             fr
         })
